Usability improvements for the run-with-gce.sh script.
This change fixes two notable usability issues with running Datalab
against a kernel gateway in a GCE VM. Specifically, with this change we will:

1. Create a separate VM for each user.
2. Enforce that the user's SSH keys are copied to that VM before trying
   to connect to it.

The first change fixes an issue where multiple users could have been
making changes to the same kernel if they were using the same kernel
gateway. This was due to the fact that the only information used to
identify a kernel in the kernel gateway was the path to and name of
the notebook. As such, if two users were running the same notebook,
they would have wound up touching the same kernel (and sharing
mutable state).

The second change fixes a first-time-use issue where the script would
not cause a user's SSH keys to be transferred to the gateway VM before
trying to connect to that VM using SSH. This meant that a user who had
connected to other VMs (using gcloud) would have been fine, but any
other user would have hit permission denied issues.